requirements.txt
# NiVi Secure Cloud Storage - Requirements
Django==5.1.7
psycopg2-binary==2.9.9
python-dotenv==1.0.1
cryptography>=41.0.0
python-social-auth>=0.3.6
social-auth-app-django>=5.0.0
python-magic>=0.4.27
python-magic-bin>=0.4.14
django-crispy-forms>=2.0
Pillow>=10.0.0